+++ telegramium-core/src/main/scala/telegramium/bots/Voice.scala
@@ -12,12 +12,14 @@ package telegramium.bots
   * @param mimeType
   *   Optional. MIME type of the file as defined by sender
   * @param fileSize
-  *   Optional. File size in bytes
+  *   Optional. File size in bytes. It can be bigger than 2&#94;31 and some programming languages may have
+  *   difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a signed 64-bit integer
+  *   or double-precision float type are safe for storing this value.
   */
 final case class Voice(
   fileId: String,
   fileUniqueId: String,
   duration: Int,
   mimeType: Option[String] = Option.empty,
-  fileSize: Option[Int] = Option.empty
+  fileSize: Option[Long] = Option.empty
 )
